+++ src/Plugins/Connect/MSFSSimConnectPlugin/src/SimVar/SimConnectFlightInfo.h
@@ -106,15 +106,19 @@ struct SimConnectFlightInfo
         flightCondition.precipitationState = toPrecipitationState(ambientPrecipState);
         flightCondition.inClouds = (ambientInCloud != 0);
 
+        // Local start time
         QTime time = QTime::fromMSecsSinceStartOfDay(localTime * 1000);
         flightCondition.startLocalDateTime.setTime(time);
         QDate date = QDate(localYear, localMonth, localDay);
         flightCondition.startLocalDateTime.setDate(date);
 
+        // Zulu start time
+        QDateTime startZuluDateTime;
         time = QTime::fromMSecsSinceStartOfDay(zuluTime * 1000);
-        flightCondition.startZuluDateTime.setTime(time);
+        startZuluDateTime.setTime(time);
         date = QDate(zuluYear, zuluMonth, zuluDay);
-        flightCondition.startZuluDateTime.setDate(date);
+        startZuluDateTime.setDate(date);
+        flightCondition.setStartZuluDateTime(startZuluDateTime);
 
         return flightCondition;
     }