+++ src/Plugins/Flight/Import/GpxImport/src/GpxParser.cpp
@@ -111,7 +111,7 @@ std::vector<FlightData> GpxParser::parseGPX() noexcept
         }
     }
     flightData.creationTime = d->firstDateTimeUtc;
-    flightData.flightCondition.startZuluDateTime = d->firstDateTimeUtc;
+    flightData.flightCondition.setStartZuluDateTime(d->firstDateTimeUtc);
     flightData.flightCondition.startLocalDateTime = d->firstDateTimeUtc.toLocalTime();
     flights.push_back(std::move(flightData));
     return flights;